+++ Engine/Models/Sector.cs
@@ -7,8 +7,8 @@
         public required TextureInfo FloorTexture { get; init; }
         public required TextureInfo CeilTexture { get; init; }
 
-        public required float Floor { get; init; }
-        public required float Ceil { get; init; }
+        public required int Floor { get; init; }
+        public required int Ceil { get; init; }
         public required RenderableWall[] Walls { get; init; }
         public required byte LightLevel { get; set; }
         public required float RotationFloor { get; set; }
